Task functionalities and assign users to task

+++ routes/api.php
@@ -2,6 +2,8 @@
 
 use Illuminate\Http\Request;
 use \App\Http\Controllers\TaskController;
+use \App\Http\Controllers\UserController;
+use \App\Http\Middleware\isAdmin;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -15,14 +17,21 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
+/*
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
+*/
 
 Route::controller(TaskController::class)->prefix('tasks')->group(function(){
-    Route::get('','getTasks');
-    Route::post('','createTask');
-    Route::put('/{task_id}','updateTask');
-    Route::delete('/{task_id}','deleteTask');
+    Route::get('','getTasks')->middleware('auth:sanctum');;
+    Route::post('','createTask')->middleware(['auth:sanctum',isAdmin::class]);;
+    Route::put('/{task_id}','updateTask')->middleware(['auth:sanctum',isAdmin::class]);;
+    Route::delete('/{task_id}','deleteTask')->middleware(['auth:sanctum',isAdmin::class]);;
 });
 
+Route::controller(UserController::class)->prefix('user')->group(function () {
+    Route::post('login','login');
+    Route::post('/register', 'register');
+    Route::post('/task/{task_id}', 'assignTask')->middleware(['auth:sanctum',isAdmin::class]);
+});
